Refactor setting modal router to separate file
Moved BasicModalSettingStack configuration from index.tsx to a new basicModalSettingRouter.ts file for better modularity and maintainability. Updated imports and references in SubSettings.tsx and index.tsx accordingly.
